requirements_web.txt
streamlit>=1.45.0
langchain-core
langchain-community
langchain-openai
chromadb
sentence-transformers
biopython
requests 